# Replace implementation of atomic addition.