@galv 2016-09-14
Note that the entire comment up there is actually the body of the commit.  Not sure whether this should be merged immediately considering that the change is small.