@danpovey 2016-09-14
Cool!  merging.